* commit 'eb1baf3a34f54361522df9dbd5200f555721eff9': (32 commits)
  quercus curl ssl HostnameVerifier issue with aws
  quercus simplexml json_encode() fix
  quercus simplexml utf8 and isset() fix
  quercus date mktime() negative values fix
  resin security doc update
  MemorySingleSignon session timeout fix
  array index issue fix
  copyright updates.
  iis bugfix
  changes.xtp fix
  quercus pdo mysql utf8 fix
  win sendfile
  win32 sendfile filename issue
  4.0.48 check-in
  add quercus support for java.sql.Timestamp2
  add quercus support for java.sql.Timestamp
  add missing mime types to app-default.xml2
  add missing mime types to app-default.xml
  add msoffice mime types to app-default.xml
  cache/database updates
  ...

# Conflicts:
#	modules/quercus/src/com/caucho/quercus/lib/db/JdbcPreparedStatementResource.java

+++ modules/resin/src/com/caucho/db/sql/DeleteQuery.java
@@ -69,35 +69,39 @@ class DeleteQuery extends Query {
     int count = 0;
     TableIterator []rows = new TableIterator[1];
 
-    try {
-      rows[0] = _table.createTableIterator();
-      context.init(xa, rows, isReadOnly());
-      
-      if (! start(rows, rows.length, context, xa)) {
-        return;
-      }
+    synchronized (_table) {
+      try {
+        rows[0] = _table.createTableIterator();
+        context.init(xa, rows, isReadOnly());
 
-      do {
-        context.lock();
-        
-        try {
-          if (isSelect(context) && rows[0].delete()) {
-            context.setRowUpdateCount(++count);
-          }
-        } finally {
-          context.unlock();
+        if (! start(rows, rows.length, context, xa)) {
+          return;
         }
-        
-        xa.commit();
-      } while (nextTuple(rows, rows.length, context, xa));
-    } catch (IOException e) {
-      throw new SQLExceptionWrapper(e);
-    } finally {
-      // autoCommitWrite must be before freeRows in case freeRows
-      // throws an exception
-      context.close();
-      
-      freeRows(rows, rows.length);
+
+        do {
+          if (! context.lock()) {
+            throw new IllegalStateException("unable to lock for delete");
+          }
+
+          try {
+            if (isSelect(context) && rows[0].delete()) {
+              context.setRowUpdateCount(++count);
+            }
+          } finally {
+            context.unlock();
+          }
+
+          xa.commit();
+        } while (nextTuple(rows, rows.length, context, xa));
+      } catch (IOException e) {
+        throw new SQLExceptionWrapper(e);
+      } finally {
+        // autoCommitWrite must be before freeRows in case freeRows
+        // throws an exception
+        context.close();
+
+        freeRows(rows, rows.length);
+      }
     }
   }
